# backend/docs/swagger.yaml
basePath: /api
definitions:
  branch.Branch:
    properties:
      capacity:
        example: 30
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Siam
        type: string
    type: object
  branch.UpdateCapacityRequest:
    properties:
      capacity:
        example: 30
        type: integer
    required:
    - capacity
    type: object
...
  scheduling.Booking:
    properties:
      branch_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      student_id:
        type: integer
      student_name:
        type: string
      subject_id:
        type: integer
      teacher_id:
        type: integer
    type: object
  scheduling.BookingAlternative:
    properties:
      branch_id:
        type: integer
      commute_minutes:
        type: integer
      end_time:
        type: string
      reasons:
        items:
          type: string
        type: array
      score:
        type: integer
      start_time:
        type: string
      subject_id:
        type: integer
      teacher_id:
        type: integer
      teacher_name:
        type: string
    type: object
  scheduling.BookingRequest:
    properties:
      branch_id:
        example: 1
        type: integer
      duration_minutes:
        example: 60
        type: integer
      preferred_slots:
        items:
          $ref: '#/definitions/shared.WeeklySlot'
        type: array
      preferred_teacher_id:
        example: 5
        type: integer
      required_gender:
        enum:
        - male
        - female
        - lgbtq+
        example: female
        type: string
      subject_id:
        example: 1
        type: integer
    required:
    - branch_id
    - preferred_slots
    - required_gender
    - subject_id
    type: object
  scheduling.BookingResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/scheduling.SlotResult'
        type: array
    type: object
  scheduling.ConfirmBookingRequest:
    properties:
      branch_id:
        example: 1
        type: integer
      end_time:
        example: "2026-06-01T10:00:00Z"
        type: string
      start_time:
        example: "2026-06-01T09:00:00Z"
        type: string
      student_id:
        example: 3
        type: integer
      subject_id:
        example: 1
        type: integer
      teacher_id:
        example: 1
        type: integer
    required:
    - branch_id
    - end_time
    - start_time
    - student_id
    - subject_id
    - teacher_id
    type: object
  scheduling.ErrorResponse:
    properties:
      error:
        example: failed to retrieve teachers
        type: string
    type: object
  scheduling.MessageResponse:
    properties:
      message:
        example: Availability saved successfully
        type: string
    type: object
  scheduling.SlotResult:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/scheduling.BookingAlternative'
        type: array
      exact_match:
        $ref: '#/definitions/scheduling.BookingAlternative'
      message:
        type: string
      slot:
        $ref: '#/definitions/shared.WeeklySlot'
    type: object
...
  shared.WeeklySlot:
    properties:
      day_of_week:
        description: 0=Sunday ... 6=Saturday (matches JavaScript Date.getDay())
        example: 0
        type: integer
      end:
        example: "10:00"
        type: string
      start:
        example: "09:00"
        type: string
    type: object
...
info:
  contact: {}
  description: API for teacher availability management and bookings
  title: Converge API
  version: "1.0"
paths:
...
  /bookings:
    get:
      description: Admin sees all bookings; a student sees only their own; a parent
        sees their students' bookings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scheduling.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List bookings (scoped by role)
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Evaluates each preferred_slot independently and returns per-slot
        results with either an exact_match or alternatives.
      parameters:
      - description: Booking request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scheduling.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Evaluate a booking request
      tags:
      - bookings
  /bookings/{id}:
    delete:
      description: Deletes a confirmed booking by ID. Returns 404 if not found.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - bookings
  /bookings/confirm:
    post:
      consumes:
      - application/json
      description: Creates the actual booking in the database after the client selects
        a result from the evaluate endpoint.
      parameters:
      - description: Booking confirmation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scheduling.ConfirmBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/scheduling.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm a booking
      tags:
      - bookings
  /branches:
    get:
      description: Returns all branches ordered by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/branch.Branch'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List branches
      tags:
      - branches
  /branches/{id}/capacity:
    patch:
      consumes:
      - application/json
      description: Sets how many bookings a branch can hold concurrently. 0 means
        unlimited/unenforced.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: New capacity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/branch.UpdateCapacityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a branch's capacity
      tags:
      - branches
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT from /login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"